requirements.txt
gradio
faster-whisper
ffmpeg-python
flask